- changed hash and hashmap classes.
-constructor now takes in a function
	that will print the value type to a string, depending on what the user
uses as a value type. although value types are stored as pointers, if
the value is not owned by the hashnode it will try to delete it, and
cause an error.

Test_HashMap now has a function to print to string an int value to pass
into the hash constructor.

TODO
each node has access to function as well as the full map. Potentially
giving each node an ability to hold different data types, if we have the
ability to set a hashnode function better.