Well I made a mess and should have made more commits and an additional branch to work in during all this updated work.
I made a mess I need to clean up, compact, refactor and automate.

Lots of new code can be put into functions or class methods.

Some class variables are no longer needed.

the use of references and deleting has to be redone as it doens't work
much.


tons of changes:
new memory manager that adds cells of memory of everthing that is
allocated with new and freed with delete.
other changes around the code but mostly spagahetti as I tried fixing a
fuew bugs that led me down a tangent. But after lots of additions and
reversals I have an existing codebase from which to continue from.

Test harnesses /unit tests still need to be created for all classes,
methods, functions, etc.
Readjustments to classes and cleanup soon to follow.