new StrObjHashMap which now inherits from generic HashMap which in turn implements a hash interface.
although not as sound and tight as it can be, hash interface and hashmap
adjusted to accomodate for different keys and values.

new strobjhashmap takes place of the old hashmap that was used with keys
as strings and values as objects or boolean numbers.

now can code an additional hashmap like expression keys and literal
values or something and less code can be written afterwards. but lots of
redundancy to already get rid of as the code is getting more and more
convoluted and thicker. Maybe it can be reworked after it is all said
and done. Although every coding session i have to step back and rework
things anyways, its getting sloppy.

not perfect and many bugs but want to commit while all example files
still work.